feat: toast refactor
