+++ ui/desktop/src/components/settings/SettingsView.tsx
@@ -15,7 +15,7 @@ import { RecentModelsRadio } from './models/RecentModels';
 import { ExtensionItem } from './extensions/ExtensionItem';
 import type { View } from '../../App';
 import { ModeSelection } from './basic/ModeSelection';
-import { ToastSuccess } from './models/toasts';
+import { toastSuccess } from '../../toasts';
 
 const EXTENSIONS_DESCRIPTION =
   'The Model Context Protocol (MCP) is a system that allows AI models to securely connect with local or remote resources using standard server setups. It works like a client-server setup and expands AI capabilities using three main components: Prompts, Resources, and Tools.';
@@ -164,7 +164,7 @@ export default function SettingsView({
     const response = await removeExtension(extensionBeingConfigured.name, true);
 
     if (response.ok) {
-      ToastSuccess({
+      toastSuccess({
         title: extensionBeingConfigured.name,
         msg: `Successfully removed extension`,
       });
