+++ ui/desktop/src/components/settings/extensions/ConfigureExtensionModal.tsx
@@ -5,8 +5,7 @@ import { Input } from '../../ui/input';
 import { FullExtensionConfig } from '../../../extensions';
 import { getApiUrl, getSecretKey } from '../../../config';
 import { addExtension } from '../../../extensions';
-import { toast } from 'react-toastify';
-import { ToastError, ToastSuccess } from '../models/toasts';
+import { toastError, toastSuccess } from '../../../toasts';
 
 interface ConfigureExtensionModalProps {
   isOpen: boolean;
@@ -71,7 +70,7 @@ export function ConfigureExtensionModal({
         throw new Error('Failed to add system configuration');
       }
 
-      ToastSuccess({
+      toastSuccess({
         title: extension.name,
         msg: `Successfully configured extension`,
       });
@@ -79,7 +78,7 @@ export function ConfigureExtensionModal({
       onClose();
     } catch (error) {
       console.error('Error configuring extension:', error);
-      ToastError({
+      toastError({
         title: extension.name,
         msg: `Failed to configure extension`,
         traceback: error.message,
