+++ ui/desktop/src/components/settings/models/utils.tsx
@@ -2,7 +2,7 @@ import { useModel } from './ModelContext'; // Import the useModel hook
 import { Model } from './ModelContext';
 import { useMemo } from 'react';
 import { gooseModels } from './GooseModels';
-import { ToastError, ToastSuccess } from './toasts';
+import { toastError, toastSuccess } from '../../../toasts';
 import { initializeSystem } from '../../../utils/providerUtils';
 import { useRecentModels } from './RecentModels';
 
@@ -32,7 +32,7 @@ export function useHandleModelSelection() {
       console.log(`[${componentName}] Switched to model: ${model.name} (${model.provider})`);
 
       // Display a success toast notification
-      ToastSuccess({
+      toastSuccess({
         title: 'Model changed',
         msg: `Switched to ${model.alias ?? model.name}`,
       });
@@ -40,7 +40,7 @@ export function useHandleModelSelection() {
       // Handle errors gracefully
       console.error(`[${componentName}] Failed to switch model:`, error);
       // Display an error toast notification
-      ToastError({
+      toastError({
         title: model.name,
         msg: `Failed to switch to model`,
         traceback: error.message,
